backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
